requirements.txt
discord.py
requests
flask
psutil
aiohttp
